requirements.txt
canonicalwebteam.flask-base==0.1.0
canonicalwebteam.templatefinder==0.1.1
canonicalwebteam.blog==2.0.9
Flask==1.0.2
